homework_6/requirements.txt
django==4.2.6
django-debug-toolbar==4.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
envparse==0.2.0
uwsgi==2.0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf_yasg==1.21.7